+++ src/main/java/me/andrz/jackson/JsonReference.java
@@ -1,264 +1,100 @@
 package me.andrz.jackson;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import com.fasterxml.jackson.core.JsonPointer;
 
-import java.io.File;
-import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.net.URL;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Set;
 
 /**
- *
+ * Represents a JSON reference string (its URI and pointer).
  */
 public class JsonReference {
 
-    private static final Logger logger = LogManager.getLogger(JsonReference.class);
+    private URI uri;
+    private JsonPointer pointer;
+    private boolean local;
+    private boolean absolute;
 
-    private ObjectMapper mapper;
-    private int maxDepth = 1;
-    private boolean stopOnCircular = true;
+    public JsonReference() {}
 
-    public JsonReference() {
-        mapper = new ObjectMapper();
-    }
+    public JsonReference(URI uri) {
 
-    public JsonReference(ObjectMapper mapper) {
-        this.mapper = mapper;
-    }
+        String fragment = uri.getFragment();
 
-    public ObjectMapper getMapper() {
-        return mapper;
-    }
+        JsonPointer pointer;
 
-    public void setMapper(ObjectMapper mapper) {
-        this.mapper = mapper;
-    }
-
-    public int getMaxDepth() {
-        return maxDepth;
-    }
-
-    public void setMaxDepth(int maxDepth) {
-        this.maxDepth = maxDepth;
-    }
-
-    public boolean isStopOnCircular() {
-        return stopOnCircular;
-    }
-
-    public void setStopOnCircular(boolean stopOnCircular) {
-        this.stopOnCircular = stopOnCircular;
-    }
-
-    public JsonNode process(File file) throws JsonReferenceException, IOException {
-        JsonContext context = new JsonContext(file);
-        process(context);
-        return context.getNode();
-    }
-
-    public JsonNode process(URL url) throws JsonReferenceException, IOException {
-        JsonContext context = new JsonContext(url);
-        process(context);
-        return context.getNode();
-    }
-
-    public JsonNode process(JsonContext context) throws JsonReferenceException, IOException {
-        JsonNode node = context.getNode();
-        process(context, node);
-        return context.getNode();
-    }
-
-    public JsonNode process(JsonContext context, Set<JsonRef> processed) throws JsonReferenceException, IOException {
-        JsonNode node = context.getNode();
-        process(context, node, processed);
-        return context.getNode();
-    }
-
-    public void process(JsonContext context, JsonNode node) throws JsonReferenceException, IOException {
-        process(context, node, null);
-    }
-
-    public void process(JsonContext context, JsonNode node, Set<JsonRef> processed) throws JsonReferenceException, IOException {
-
-        if (node == null) {
-            return;
-        }
-
-        if (processed == null) {
-            processed = new HashSet<JsonRef>();
-        }
-
-        logger.debug("processed: " + processed);
-
-        if (processed.size() >= maxDepth) {
-            return;
-        }
-
-        if (node.isArray()) {
-            ArrayNode arrayNode = (ArrayNode) node;
-            Iterator<JsonNode> elements = arrayNode.elements();
-            int i = 0;
-            while (elements.hasNext()) {
-                JsonNode subNode = elements.next();
-
-                logger.debug("i=" + i);
-
-                if (JsonRefNode.is(subNode)) {
-
-                    JsonNode replacementNode = getReplacementNode(subNode, context, processed);
-                    if (replacementNode == null) continue;
-
-                    logger.debug("replacing " + "subNode" + " with " + replacementNode);
-                    arrayNode.set(i, replacementNode);
-                    ++i;
-                }
-                else {
-                    process(context, subNode, processed);
-                }
-            }
-        }
-        else if (node.isObject()) {
-            ObjectNode objectNode = (ObjectNode) node;
-
-            Iterator<Map.Entry<String, JsonNode>> fields = objectNode.fields();
-            while (fields.hasNext()) {
-                Map.Entry<String, JsonNode> field = fields.next();
-                String key = field.getKey();
-                JsonNode subNode = field.getValue();
-
-                logger.debug("key=" + key);
-
-                if (JsonRefNode.is(subNode)) {
-
-                    JsonNode replacementNode = getReplacementNode(subNode, context, processed);
-                    if (replacementNode == null) continue;
-
-                    logger.debug("replacing " + "subNode" + " with " + replacementNode);
-                    objectNode.set(key, replacementNode);
-                }
-                else {
-                    process(context, subNode, processed);
-                }
-            }
-        }
-    }
-
-    public JsonNode getReplacementNode(JsonNode subNode, JsonContext context, Set<JsonRef> processed) throws JsonReferenceException, IOException {
-
-        JsonRef ref = getJsonRefForJsonNode(subNode);
-        JsonRef absRef = getAbsoluteRef(ref, context);
-        if (stopOnCircular && processed.contains(absRef)) {
-            logger.debug("skipping on ref: " + absRef);
-            return null;
-        }
-        // add ref to processed set for detection loop in recursion
-        processed.add(absRef);
-
-        JsonContext referencedContext = resolveFromContextToContext(ref, context);
-
-        // recurse
-        process(referencedContext, processed);
-
-        // after recursing, remove ref from processed set for next iteration
-        processed.remove(absRef);
-
-        JsonNode replacementNode = referencedContext.getNode();
-
-        return replacementNode;
-    }
-
-    public JsonRef getAbsoluteRef(JsonRef ref, JsonContext context) throws JsonReferenceException {
-        // TODO: More robust URL building.
-        String newRefString = context.getUrl() + "#" + ref.getPointer().toString();
-        JsonRef newRef = JsonRef.fromString(newRefString);
-        return newRef;
-    }
-
-    public JsonRef getJsonRefForJsonNode(JsonNode node) throws JsonReferenceException {
-        JsonRefNode refNode = new JsonRefNode(node);
-        String refString = refNode.getRefString();
-        JsonRef ref = JsonRef.fromString(refString);
-        return ref;
-    }
-
-    public JsonContext resolveFromContextToContext(JsonRef ref, JsonContext context) throws IOException, JsonReferenceException {
-
-        JsonContext referencedContext;
-        JsonNode referencedNode;
-
-        URL absoluteReferencedUrl;
-        URI refUri = ref.getUri();
-
-        logger.debug("dereferencing " + ref);
-
-        if (ref.isLocal()) {
-            absoluteReferencedUrl = context.getUrl();
-            JsonNode clone = new ObjectMapper().readTree(context.getNode().traverse());
-            referencedNode = clone.at(ref.getPointer());
-        }
-        else if (ref.isAbsolute()) {
-            absoluteReferencedUrl = refUri.toURL();
-            referencedNode = read(absoluteReferencedUrl).at(ref.getPointer());
+        if (fragment == null || fragment.isEmpty()) {
+            pointer = JsonPointer.compile(null);
         }
         else {
-            URL contextUrl = context.getUrl();
-            try {
-                absoluteReferencedUrl = contextUrl.toURI().resolve(refUri).toURL();
-            } catch (URISyntaxException e) {
-                throw new JsonReferenceException("Invalid URI for context URL: " + contextUrl);
-            }
-            referencedNode = read(absoluteReferencedUrl).at(ref.getPointer());
+            pointer = JsonPointer.compile(fragment);
         }
 
-        referencedContext = new JsonContext();
-        referencedContext.setUrl(absoluteReferencedUrl);
-        referencedContext.setNode(referencedNode);
+        /*
+         * Remove any extraneous path segments, especially to make semantically empty paths really empty.
+         */
+        uri = uri.normalize();
 
-        referencedContext.setUrl(absoluteReferencedUrl);
+        absolute = uri.isAbsolute();
+        local = ! uri.isAbsolute() && "".equals(uri.getPath());
 
-        return referencedContext;
+        this.uri = uri;
+        this.pointer = pointer;
     }
 
-    /**
-     * Resolve with defaults.
-     * - Assumes ref points to absolute URL.
-     *
-     * @param ref
-     * @return
-     * @throws IOException
-     */
-    public JsonNode get(JsonRef ref) throws IOException {
-        JsonNode referencedNode;
-
-        URI refUri = ref.getUri();
-        URL url = refUri.toURL();
-
-        referencedNode = read(url).at(ref.getPointer());
-
-        return referencedNode;
+    public static JsonReference fromString(String string) throws JsonReferenceException {
+        try {
+            return fromURI(new URI(string));
+        } catch (URISyntaxException e) {
+            throw new JsonReferenceException("Invalid URI: " + string);
+        }
     }
 
-    public JsonNode read(URL url) throws IOException {
-        return mapper.readTree(url);
+    public static JsonReference fromURI(URI uri) {
+        return new JsonReference(uri);
     }
 
-    public JsonNode read(File file) throws IOException {
-        return mapper.readTree(file);
+    public boolean isLocal() {
+        return local;
     }
 
-    public JsonNode readFile(String fileString) throws IOException {
-        File file = new File(fileString);
-        return read(file);
+    public boolean isAbsolute() {
+        return absolute;
+    }
+
+    public URI getUri() {
+        return uri;
+    }
+
+    public JsonPointer getPointer() {
+        return pointer;
+    }
+
+    public String toString() {
+        return uri.toString();
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        JsonReference jsonRef = (JsonReference) o;
+
+        if (uri != null ? !uri.equals(jsonRef.uri) : jsonRef.uri != null) return false;
+
+        return true;
+    }
+
+    @Override
+    public int hashCode() {
+        return uri != null ? uri.hashCode() : 0;
+    }
+
+    @Override
+    public JsonReference clone() {
+        JsonReference ref = new JsonReference(uri);
+        return ref;
     }
 
 }
